database_schemas/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sentence-transformers>=3.0.0
numpy>=1.23
python-dateutil>=2.8.2
